Separated extension.ts into different files
- detectFile for detection algorithms for file formats
- errorPosition for helper functions that find positions/indices in the document
- parseMarkdown and parseYaml, in src/parsers folder, for parsing prompt markdown and treatments yaml respectively
- extension.ts is the main function using these helper functions

Merged in broken yaml files but have not merged in the test cases for these yet
