+++ src/test/suite/fixtures/missingName.md
@@ -1,8 +1,5 @@
 ---
-name: src/test/suite/fixtures/missingName.md
-type: openRespons
-notes:
-rows: d
+type: noResponse
 ---
 
 Please describe the chair you are sitting in.
